Разделяет код на отдельные ES2015-модули, подключает главный модуль в index.html как скрипт-модуль
